Completed the toast popup and sticky header

+++ 6_advanced/vuex-token/src/views/MainView.vue
@@ -1,15 +1,25 @@
 <template>
   <div>
-    <button @click="$router.push('new')">add a new post</button>
-    <div>
+    <div class="list-container">
       <ul>
         <li v-for="item in postItems" :key="item._id">
-          <router-link :to="`/post/${item._id}`">
+          <div class="post-title">
+            <router-link :to="`/post/${item._id}`">{{ item.title }}</router-link>
+          </div>
+          <div class="post-contents">
+            {{ item.contents }}
+          </div>
+          <div class="post-time">
+            {{ item.createdAt }}
+            <i class="icon ion-md-create"></i>
+            <i class="icon ion-md-trash" @click="removePost(item._id)"></i>
+          </div>
+          <!-- <router-link :to="`/post/${item._id}`">
             {{ item.title }}
           </router-link>
           <button @click="removePost(item._id)">
             delete
-          </button>
+          </button> -->
         </li>
       </ul>
     </div>
@@ -18,6 +28,7 @@
 
 <script>
 import { fetchPosts, deletePostById } from '../api/index.js';
+import bus from '../utils/bus.js';
 
 export default {
   data() {
@@ -37,8 +48,12 @@ export default {
     },
     async removePost(id) {
       try {
-        await deletePostById(id);
-        await this.fetchData();
+        if (confirm('Delete it?')) {
+          const response = await deletePostById(id);
+          console.log(response);
+          await this.fetchData();
+          bus.$emit('show:toast', `${response.data.title} was deleted`);
+        }
       } catch (error) {
         console.log(error);
       }
@@ -50,6 +65,50 @@ export default {
 }
 </script>
 
-<style>
-
+<style scoped>
+.list-container.sticky {
+  margin-top: 76px;
+}
+ul {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+ul>li {
+  width: 450px;
+  height: 250px;
+  border: 1px solid #dae1e7;
+  border-radius: 35px;
+  box-shadow: 0 2px 4px 0 rgba(0,0,0,.1);
+  margin: 1rem 0;
+  padding: 1rem;
+}
+.post-title {
+  text-align: center;
+  font-size: 1.8rem;
+  font-weight: 600;
+  margin-bottom: 0.5rem;
+}
+.post-contents {
+  height: 73%;
+}
+.post-time {
+  font-size: 1rem;
+  text-align: right;
+}
+.icon {
+  font-size: 1.3rem;
+  cursor: pointer;
+  color: #364f6b;
+  padding-right: 0.4rem;
+}
+.icon:hover {
+  color: #3fc1c9;
+}
+.icon:active {
+  color: #fc5185;
+}
+.ion-md-create {
+  padding-left: 0.1rem;
+}
 </style>
